Added Images, Transitions
Basic images added to the sorting page. I’ve attempted to add the
beginning of a new system for the sorting pages that incorporates two
pages, both included in sorting.html. The first page is for selecting
the customer name and color card that will be associated with the
customer’s laundry throughout the process (still haven’t added that
part in). The second page will be where the employee will perform the
clothing counts.