Initial back-end work. UI Improvements
I worked for a while on it, but I couldn’t get sqlite to work with the
unc cs department servers, so I have been testing this locally with
XAMPP.

This commit contains the beginnings of the backend. I have written an
ORM layer for the Customer and Employee database tables, and have
initial work on the Laundry table. The structure of the database is
subject to change at this point however.